Move test_helpers to testing/. (#1321)
