+++ cartographer/pose_graph/constraint/cost_function/relative_pose_cost_2d_test.cc
@@ -17,7 +17,7 @@
 #include "cartographer/pose_graph/constraint/cost_function/relative_pose_cost_2d.h"
 
 #include "cartographer/common/make_unique.h"
-#include "cartographer/pose_graph/internal/testing/test_helpers.h"
+#include "cartographer/testing/test_helpers.h"
 #include "ceres/gradient_checker.h"
 
 namespace cartographer {
