+++ cartographer/pose_graph/node/node_id_test.cc
@@ -18,7 +18,7 @@
 
 #include <sstream>
 
-#include "cartographer/pose_graph/internal/testing/test_helpers.h"
+#include "cartographer/testing/test_helpers.h"
 
 namespace cartographer {
 namespace pose_graph {
