+++ cartographer/pose_graph/optimizer/ceres_optimizer_test.cc
@@ -18,7 +18,7 @@
 
 #include "cartographer/common/make_unique.h"
 #include "cartographer/pose_graph/constraint/relative_pose_constraint_2d.h"
-#include "cartographer/pose_graph/internal/testing/test_helpers.h"
+#include "cartographer/testing/test_helpers.h"
 
 namespace cartographer {
 namespace pose_graph {
